Refactored Users to be User

+++ src/main/java/com/brunojustino/restapi/repository/JournalRepository.java
@@ -3,5 +3,5 @@ package com.brunojustino.restapi.repository;
 import com.brunojustino.restapi.entity.Journal;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface JournalRepository extends JpaRepository<Journal, Integer> {
+public interface JournalRepository extends JpaRepository<Journal, Long> {
 }
